pyproject.toml
[project]
name = "Flask_Authentication_API"
requires-python = ">=3.13"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
dependencies = [
    "cryptography>=46.0.5",
    "dotenv>=0.9.9",
    "flask>=3.1.2",
    "flask-cors>=6.0.2",
    "flask-login>=0.6.3",
    "flask-migrate>=4.1.0",
    "flask-sqlalchemy>=3.1.1",
    "pandas>=3.0.0",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pyjwt>=2.11.0",
    "pytest>=9.0.2",
    "requests>=2.32.5",
]
[tool.testing.pytest.ini_options]
pythonpath = ["./src", "./tests"]

[tool.ty.rules]
unsupported-base = "ignore"